Settings: Update maintainer name
For unofficial/official builds, it can be set to "Unspecified" or the Rom's lead name (in this example, Aryan Gupta) till and until the maintainer overrides it. In this scenario, it's best to utilise the ROM's lead name so that even if it's not overwritten, it displays some name (the leads name) rather than none.